\name{affinity.spline.coefs}
\docType{data}
\alias{affinity.spline.coefs}
\title{Spline coefficients for estimation of affinity from probe sequence}
\description{Spline coefficients for estimation of affinity from probe sequence}
\usage{data(affinity.spline.coefs)}
%\format{}
%\source{}
%\url{}
\seealso{\code{\link{compute.affinities}}}
\keyword{datasets}
